import-url: correcting the split dir - output in utils.resolve_paths
