+++ tests/dir_helpers.py
@@ -382,18 +382,3 @@ def git_dir(make_tmp_dir):
     path = make_tmp_dir("git-erepo", scm=True)
     path.scm.commit("init repo")
     return path
-
-
-def load_gdrive_credentials(root_dir):
-    current_dir = pathlib.Path(__file__).parent
-    gdrive_credentials = current_dir.joinpath("gdrive-user-credentials.json")
-
-    if gdrive_credentials.exists():
-        import shutil
-
-        inner_tmp = pathlib.Path(root_dir).joinpath(".dvc", "tmp")
-        inner_tmp.mkdir(exist_ok=True)
-        shutil.copy(gdrive_credentials, inner_tmp)
-        return True
-
-    return False
