+++ tests/utils/__init__.py
@@ -1,4 +1,5 @@
 import os
+import pathlib
 from contextlib import contextmanager
 
 from dvc.scm import Git
@@ -21,3 +22,18 @@ def cd(newdir):
 
 def to_posixpath(path):
     return path.replace("\\", "/")
+
+
+def load_gdrive_credentials(root_dir):
+    current_dir = pathlib.Path(__file__).parent
+    gdrive_credentials = current_dir.joinpath("gdrive-user-credentials.json")
+
+    if gdrive_credentials.exists():
+        import shutil
+
+        inner_tmp = pathlib.Path(root_dir).joinpath(".dvc", "tmp")
+        inner_tmp.mkdir(exist_ok=True)
+        shutil.copy(gdrive_credentials, inner_tmp)
+        return True
+
+    return False
